// src/types/products.ts
export enum Condition {
	NEW = "New",
	REFURBISHED = "Refurbished",
	USED = "Used",
}

export enum EnergyEfficiencyClass {
	A_PLUS_PLUS_PLUS = "A+++",
	A_PLUS_PLUS = "A++",
	A_PLUS = "A+",
	A = "A",
	B = "B",
	C = "C",
	D = "D",
	E = "E",
	F = "F",
	G = "G",
}

export enum CurrencyCode {
	AED = "AED-United Arab Emirates Dirham",
	AFN = "AFN-Afghan Afghani",
	ALL = "ALL-Albanian Lek",
	AMD = "AMD-Armenian Dram",
	ANG = "ANG-Netherlands Antillean Guilder",
	AOA = "AOA-Angolan Kwanza",
	ARS = "ARS-Argentine Peso",
	AUD = "AUD-Australian Dollar",
	AWG = "AWG-Aruban Florin",
	AZN = "AZN-Azerbaijani Manat",
	BAM = "BAM-Bosnia-Herzegovina Convertible Mark",
	BBD = "BBD-Barbadian Dollar",
	BDT = "BDT-Bangladeshi Taka",
	BGN = "BGN-Bulgarian Lev",
	BHD = "BHD-Bahraini Dinar",
	BIF = "BIF-Burundian Franc",
	BMD = "BMD-Bermudan Dollar",
	BND = "BND-Brunei Dollar",
	BOB = "BOB-Bolivian Boliviano",
	BRL = "BRL-Brazilian Real",
	BSD = "BSD-Bahamian Dollar",
	BTN = "BTN-Bhutanese Ngultrum",
	BWP = "BWP-Botswanan Pula",
	BYN = "BYN-Belarusian Ruble",
	BZD = "BZD-Belize Dollar",
	CAD = "CAD-Canadian Dollar",
	CDF = "CDF-Congolese Franc",
	CHF = "CHF-Swiss Franc",
	CLP = "CLP-Chilean Peso",
	CNY = "CNY-Chinese Yuan",
	COP = "COP-Colombian Peso",
	CRC = "CRC-Costa Rican Colón",
	CUP = "CUP-Cuban Peso",
	CVE = "CVE-Cape Verdean Escudo",
	CZK = "CZK-Czech Koruna",
	DJF = "DJF-Djiboutian Franc",
	DKK = "DKK-Danish Krone",
	DOP = "DOP-Dominican Peso",
	DZD = "DZD-Algerian Dinar",
	EGP = "EGP-Egyptian Pound",
	ERN = "ERN-Eritrean Nakfa",
	ETB = "ETB-Ethiopian Birr",
	EUR = "EUR-Euro",
	FJD = "FJD-Fijian Dollar",
	FKP = "FKP-Falkland Islands Pound",
	GBP = "GBP-British Pound",
	GEL = "GEL-Georgian Lari",
	GHS = "GHS-Ghanaian Cedi",
	GIP = "GIP-Gibraltar Pound",
	GMD = "GMD-Gambian Dalasi",
	GNF = "GNF-Guinean Franc",
	GTQ = "GTQ-Guatemalan Quetzal",
	GYD = "GYD-Guyanaese Dollar",
	HKD = "HKD-Hong Kong Dollar",
	HRK = "HRK-Croatian Kuna",
	HTG = "HTG-Haitian Gourde",
	HUF = "HUF-Hungarian Forint",
	IDR = "IDR-Indonesian Rupiah",
	ILS = "ILS-Israeli New Shekel",
	INR = "INR-Indian Rupee",
	IQD = "IQD-Iraqi Dinar",
	IRR = "IRR-Iranian Rial",
	ISK = "ISK-Icelandic Króna",
	JMD = "JMD-Jamaican Dollar",
	JOD = "JOD-Jordanian Dinar",
	JPY = "JPY-Japanese Yen",
	KES = "KES-Kenyan Shilling",
	KGS = "KGS-Kyrgystani Som",
	KHR = "KHR-Cambodian Riel",
	KMF = "KMF-Comorian Franc",
	KRW = "KRW-South Korean Won",
	KWD = "KWD-Kuwaiti Dinar",
	KYD = "KYD-Cayman Islands Dollar",
	KZT = "KZT-Kazakhstani Tenge",
	LAK = "LAK-Laotian Kip",
	LBP = "LBP-Lebanese Pound",
	LKR = "LKR-Sri Lankan Rupee",
	LRD = "LRD-Liberian Dollar",
	LSL = "LSL-Lesotho Loti",
	LYD = "LYD-Libyan Dinar",
	MAD = "MAD-Moroccan Dirham",
	MDL = "MDL-Moldovan Leu",
	MGA = "MGA-Malagasy Ariary",
	MKD = "MKD-Macedonian Denar",
	MMK = "MMK-Myanmar Kyat",
	MNT = "MNT-Mongolian Tugrik",
	MOP = "MOP-Macanese Pataca",
	MRO = "MRO-Mauritanian Ouguiya (1973–2017)",
	MUR = "MUR-Mauritian Rupee",
	MVR = "MVR-Maldivian Rufiyaa",
	MWK = "MWK-Malawian Kwacha",
	MXN = "MXN-Mexican Peso",
	MYR = "MYR-Malaysian Ringgit",
	MZN = "MZN-Mozambican Metical",
	NAD = "NAD-Namibian Dollar",
	NGN = "NGN-Nigerian Naira",
	NIO = "NIO-Nicaraguan Córdoba",
	NOK = "NOK-Norwegian Krone",
	NPR = "NPR-Nepalese Rupee",
	NZD = "NZD-New Zealand Dollar",
	OMR = "OMR-Omani Rial",
	PAB = "PAB-Panamanian Balboa",
	PEN = "PEN-Peruvian Sol",
	PGK = "PGK-Papua New Guinean Kina",
	PHP = "PHP-Philippine Piso",
	PKR = "PKR-Pakistani Rupee",
	PLN = "PLN-Polish Zloty",
	PYG = "PYG-Paraguayan Guarani",
	QAR = "QAR-Qatari Rial",
	RON = "RON-Romanian Leu",
	RSD = "RSD-Serbian Dinar",
	RUB = "RUB-Russian Ruble",
	RWF = "RWF-Rwandan Franc",
	SAR = "SAR-Saudi Riyal",
	SBD = "SBD-Solomon Islands Dollar",
	SCR = "SCR-Seychellois Rupee",
	SDG = "SDG-Sudanese Pound",
	SEK = "SEK-Swedish Krona",
	SGD = "SGD-Singapore Dollar",
	SHP = "SHP-St. Helena Pound",
	SLL = "SLL-Sierra Leonean Leone",
	SOS = "SOS-Somali Shilling",
	SRD = "SRD-Surinamese Dollar",
	SSP = "SSP-South Sudanese Pound",
	STD = "STD-São Tomé & Príncipe Dobra (1977–2017)",
	SYP = "SYP-Syrian Pound",
	SZL = "SZL-Swazi Lilangeni",
	THB = "THB-Thai Baht",
	TJS = "TJS-Tajikistani Somoni",
	TMT = "TMT-Turkmenistani Manat",
	TND = "TND-Tunisian Dinar",
	TOP = "TOP-Tongan Paʻanga",
	TRY = "TRY-Turkish Lira",
	TTD = "TTD-Trinidad & Tobago Dollar",
	TWD = "TWD-New Taiwan Dollar",
	TZS = "TZS-Tanzanian Shilling",
	UAH = "UAH-Ukrainian Hryvnia",
	UGX = "UGX-Ugandan Shilling",
	USD = "USD-US Dollar",
	UYU = "UYU-Uruguayan Peso",
	UZS = "UZS-Uzbekistani Som",
	VEF = "VEF-Venezuelan Bolívar (2008–2018)",
	VND = "VND-Vietnamese Dong",
	VUV = "VUV-Vanuatu Vatu",
	WST = "WST-Samoan Tala",
	XAF = "XAF-Central African CFA Franc",
	XCD = "XCD-East Caribbean Dollar",
	XOF = "XOF-West African CFA Franc",
	XPF = "XPF-CFP Franc",
	YER = "YER-Yemeni Rial",
	ZAR = "ZAR-South African Rand",
	ZMW = "ZMW-Zambian Kwacha",
}

export interface Price {
	value?: number,
	currencyCode: CurrencyCode,
}

export enum Period {
	MONTH = "Month",
	YEAR = "Year",
}

export interface InstallmentPlan {
	numberOfPeriods: number,
	price: Price,
	period: Period,
}

export interface LoyaltyPoints {
	loyaltyProgramName?: string,
	numberOfPoints: number,
	ratioToCurrency?: number,
}

export enum MaximumEnergyEfficiencyClass {
	A_PLUS_PLUS_PLUS = "A+++",
	A_PLUS_PLUS = "A++",
	A_PLUS = "A+",
	A = "A",
	B = "B",
	C = "C",
	D = "D",
	E = "E",
	F = "F",
	G = "G",
}

export enum MinimumEnergyEfficiencyClass {
	A_PLUS_PLUS_PLUS = "A+++",
	A_PLUS_PLUS = "A++",
	A_PLUS = "A+",
	A = "A",
	B = "B",
	C = "C",
	D = "D",
	E = "E",
	F = "F",
	G = "G",
}

export interface EntityReference {
	entityId: string,
	name: string,
}

export interface ImageThumbnail {
	url: string,
	width: number,
	height: number,
}

export interface Image {
	url: string,
	width: number,
	height: number,
	thumbnails?: ImageThumbnail[],
	alternateText?: string,
}

export interface ComplexImage {
	image: Image,
	details?: string,
	description?: string,
	clickthroughUrl?: string,
}

export interface ProductionCost {
	value?: number,
	currencyCode: CurrencyCode,
}

export interface SalePrice {
	value?: number,
	currencyCode: CurrencyCode,
}

export interface Sales {
	salePrice?: SalePrice,
	saleStartDate?: string,
	saleEndDate?: string,
}

export enum SalesChannel {
	LOCAL = "Local",
	ONLINE = "Online",
}

export interface ShippingHeight {
	unit: string,
	value: number,
}

export interface ShippingLength {
	unit: string,
	value: number,
}

export interface ShippingWeight {
	unit: string,
	value: number,
}

export interface ShippingWidth {
	unit: string,
	value: number,
}

export enum SizeSystem {
	US = "US",
	UK = "UK",
	EU = "EU",
	DE = "DE",
	FR = "FR",
	JP = "JP",
	CN = "CN",
	IT = "IT",
	BR = "BR",
	MEX = "MEX",
	AU = "AU",
}

export enum SizeType {
	REGULAR = "Regular",
	PETITE = "Petite",
	MATERNITY = "Maternity",
	BIG = "Big",
	TALL = "Tall",
	PLUS = "Plus",
}

export enum StockStatus {
	IN_STOCK = "In Stock",
	OUT_OF_STOCK = "Out of Stock",
	BACKORDERED = "Backordered",
	PRE_ORDER = "Pre-Order",
}

export interface SubscriptionPlan {
	numberOfPeriods: number,
	price: Price,
	period: Period,
}

export enum TargetAgeGroup {
	NEWBORN = "Newborn",
	INFANT = "Infant",
	TODDLER = "Toddler",
	KIDS = "Kids",
	ADULT = "Adult",
}

export enum TargetGender {
	MALE = "Male",
	FEMALE = "Female",
	UNISEX = "Unisex",
}

export interface UnitPricingBaseMeasure {
	unit: string,
	value: number,
}

export interface UnitPricingMeasure {
	unit: string,
	value: number,
}

export interface ComplexVideo {
	url: string,
	video?: string,
	description?: string,
}

export default interface Product {
	availabilityDate?: string,
	brand?: string,
	bundle?: boolean,
	color?: string,
	condition?: Condition,
	energyEfficiencyClass?: EnergyEfficiencyClass,
	expirationDate?: string,
	gtin?: string,
	includesAdultContent?: boolean,
	installmentPlan?: InstallmentPlan,
	inventoryQuantity?: number,
	landingPageUrl?: string,
	loyaltyPoints?: LoyaltyPoints,
	material?: string,
	maximumEnergyEfficiencyClass?: MaximumEnergyEfficiencyClass,
	minimumEnergyEfficiencyClass?: MinimumEnergyEfficiencyClass,
	mobileOptimizedLandingPage?: string,
	mpn?: string,
	nudgeEnabled?: boolean,
	numberOfItemsInPack?: number,
	parentProduct?: EntityReference,
	pattern?: string,
	price?: Price,
	primaryConversationContact?: any,
	primaryPhoto?: ComplexImage,
	productHighlights?: string[],
	productionCost?: ProductionCost,
	richTextDescriptionV2?: any,
	sales?: Sales[],
	salesChannel?: SalesChannel,
	shippingHeight?: ShippingHeight,
	shippingLength?: ShippingLength,
	shippingWeight?: ShippingWeight,
	shippingWidth?: ShippingWidth,
	size?: string,
	sizeSystem?: SizeSystem,
	sizeType?: SizeType[],
	sku?: string,
	slug?: string,
	stockStatus?: StockStatus,
	subscriptionPlan?: SubscriptionPlan,
	targetAgeGroup?: TargetAgeGroup,
	targetGender?: TargetGender,
	unitPricingBaseMeasure?: UnitPricingBaseMeasure,
	unitPricingMeasure?: UnitPricingMeasure,
	name: string,
	c_actualProductDesignOrColorMayVary?: boolean,
	c_age?: string[],
	c_apparelFit?: string[],
	c_apparelSize?: string[],
	c_assortmentByDesignColor?: string[],
	c_assortmentDetails?: string[],
	c_availableInStoreOnly?: boolean,
	c_balancedChoices?: string[],
	c_batteriesIncluded?: boolean,
	c_batterySize?: string[],
	c_beddingSize?: string[],
	c_beverageSize?: string[],
	c_bpaFree?: boolean,
	c_brand?: string[],
	c_caseQuantity?: string[],
	c_cCategory?: string,
	c_collection?: string[],
	c_colors?: string[],
	c_dietaryNeeds?: string[],
	dm_directoryParents?: EntityReference[],
	c_dishWasherSafe?: boolean,
	c_dryCleanOnly?: boolean,
	c_finish?: string[],
	c_flavor?: string[],
	c_foodAllergens?: string[],
	c_licensedProduct?: boolean,
	c_limitedQuantities?: boolean,
	c_madeInUsa?: boolean,
	c_material?: string[],
	c_microwaveSafe?: boolean,
	c_pattern?: string[],
	c_pet?: string[],
	c_productCategory1?: string,
	c_productCategory2?: string,
	c_productCategory3?: string,
	c_requiresBatteries?: boolean,
	c_scent?: string[],
	c_shape?: string[],
	c_shoeSize?: string[],
	c_sockSize?: string[],
	c_testedFoodSafe?: boolean,
	c_valanceType?: string[],
	c_windowTreatmentType?: string[],
	c_wineType?: string[],
	c_wineVarietal?: string[],
	photoGallery?: ComplexImage[],
	keywords?: string[],
	id: string,
	timezone?: any,
	videos?: ComplexVideo[],
}